Rename `ActionsNext` to `Actions`
